+++ grains/src/grains.h
@@ -1,5 +1,5 @@
 #ifndef GRAINS_H
 #define GRAINS_H
-unsigned long square(const unsigned int n);
-unsigned long total();
+unsigned long long square(const unsigned int n);
+unsigned long long total();
 #endif
